ref(ngrx-ducks): move createDuck to own file
