+++ ngrx/ducks/src/lib/ducks/index.ts
@@ -1,0 +1,1 @@
+export * from './create-duck';
